Move 2 methods in session/rbczmq common with ffi_rzmq to session/base
