+++ lib/iruby/session/rbczmq.rb
@@ -1,5 +1,9 @@
+require 'rbczmq'
+require_relative 'base.rb'
+
 module IRuby
   class Session
+    include SessionBase
     DELIM = '<IDS|MSG>'
 
     def initialize(config)
@@ -63,36 +67,6 @@ module IRuby
       frames = ([*idents].compact.map(&:to_s) << DELIM << sign(msg)) + msg
       IRuby.logger.debug "Sent #{frames.inspect}"
       ZMQ::Message(*frames)
-    end
-
-    def unserialize(msg)
-      raise 'no message received' unless msg
-      frames = msg.to_a.map(&:to_s)
-      IRuby.logger.debug "Received #{frames.inspect}"
-
-      i = frames.index(DELIM)
-      idents, msg_list = frames[0..i-1], frames[i+1..-1]
-
-      minlen = 5
-      raise 'malformed message, must have at least #{minlen} elements' unless msg_list.length >= minlen
-      s, header, parent_header, metadata, content, buffers = *msg_list
-      raise 'Invalid signature' unless s == sign(msg_list[1..-1])
-      {
-        idents:        idents,
-        header:        MultiJson.load(header),
-        parent_header: MultiJson.load(parent_header),
-        metadata:      MultiJson.load(metadata),
-        content:       MultiJson.load(content),
-        buffers:       buffers
-      }
-    end
-
-    # Sign using HMAC
-    def sign(list)
-      return '' unless @hmac
-      @hmac.reset
-      list.each {|m| @hmac.update(m) }
-      @hmac.hexdigest
     end
   end
 end
